requirements.txt
mysql-connector
numpy
pandas
matplotlib
